Add files via upload
PCOS Prediction Models
Introduction
Polycystic Ovary Syndrome (PCOS) is a common endocrine disorder affecting approximately 6–10% of women of reproductive age. It has a broad spectrum of manifestations, ranging from irregular menstrual cycles to insulin resistance and infertility. It is a major health concern, and timely prediction and diagnosis can greatly help in managing the symptoms and improving the quality of life for affected individuals.

Why Is This Important?
September is PCOS Awareness Month, a time to draw attention to this disorder that many women struggle with but that often remains underdiagnosed. Through this project, I aim to contribute to the awareness and understanding of PCOS and provide a set of machine learning models that may assist in its prediction and diagnosis.

Models Used
We employed a variety of machine learning algorithms to predict the occurrence of PCOS based on various clinical and demographic features. The models include:

Logistic Regression
Naive Bayes
Decision Trees
K-Nearest Neighbors (KNN)
CatBoost
Random Forest

Results
Prediction accuracy, precision, and other evaluation metrics vary across models. Detailed results and analysis are provided in the respective Jupyter notebooks.

the invaluable datasets provided by Kaggle.